requirements/requirements.txt
boto3==1.35.71
celery==5.4.0
celery[redis]==5.4.0
click==8.1.3
dj-database-url==0.5.0
django==5.0.6
django-browser-reload==1.6.0
django-celery-beat==2.6.0
django-compressor==4.1
django-cors-headers==3.13.0
django-countries==7.5
django-debug-toolbar
django-hijack==3.5.3
django-json-widget==2.0.1
django-jsonform==2.22.0
django-opensearch-dsl==0.6.2
django-phonenumber-field==7.0.0
django-storages==1.13.1
django-tailwind==3.8.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
factory-boy==3.2.1
google-auth==2.16.0
gunicorn==20.0.4
phonenumbers==8.13.0
psycopg2-binary
python-dotenv==0.15.0
sentry-sdk==2.18.0
serpy==0.3.1
shortuuid==1.0.11
stripe==4.2.0
whitenoise==6.2.0